refactor: move brc home page into sites/brc-analytics (#1637)
